Bug 635975 - Older jetpack addons cause tabs to become totally disorganized in Panorama [r=ian]
UltraBlame original commit: 9ef10b9170bc5c9d95648a5e334e02f7387e28ef

+++ browser/base/content/tabview/tabitems.js
@@ -857,7 +857,9 @@ let TabItems = {
       if (tab.ownerDocument.defaultView != gWindow || tab.pinned)
         return;
 
-      self.unlink(tab);
+      
+      if (!UI.isDOMWindowClosing)
+        self.unlink(tab);
     }
     for (let name in this._eventListeners) {
       AllTabs.register(name, this._eventListeners[name]);
